requirements.txt
pyspark
kafka-python
cassandra-driver
streamlit
faker
